refactored file structure for lab7
