+++ app/__init__.py
@@ -1,5 +1,20 @@
 from flask import Flask
+from flask_login import LoginManager
+from flask_migrate import Migrate
+from flask_sqlalchemy import SQLAlchemy
+
+from config import *
+
 
 app = Flask(__name__)
+app.config['SQLALCHEMY_DATABASE_URI'] = SQLALCHEMY_DATABASE_URI
+app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = SQLALCHEMY_TRACK_MODIFICATIONS
+db = SQLAlchemy(app)
+app.config['SECRET_KEY'] = SECRET_KEY
+asd = Migrate(app, db)
+
+login_manager = LoginManager(app)
+login_manager.login_view = 'login'
 
 from app import views
+
