+++ app/models.py
@@ -1,0 +1,22 @@
+from flask_login import UserMixin
+
+from app import db
+
+
+class Todo(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    title = db.Column(db.String(100), nullable=False)
+    description = db.Column(db.String(200))
+
+    def __repr__(self):
+        return f'<Todo {self.title}>'
+
+
+class Feedback(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(50), nullable=False)
+    comment = db.Column(db.Text, nullable=False)
+
+
+class User(UserMixin):
+    pass
