+++ config.py
@@ -1,0 +1,3 @@
+SQLALCHEMY_DATABASE_URI = 'sqlite:///feedbacks.db'
+SECRET_KEY = "asdasd"
+SQLALCHEMY_TRACK_MODIFICATIONS = False
